+++ rust/operator-binary/src/discovery.rs
@@ -1,24 +1,23 @@
-use crate::hbase_controller::build_recommended_labels;
-use stackable_hbase_crd::{HbaseCluster, HbaseRole, HBASE_SITE_XML, HBASE_ZOOKEEPER_QUORUM};
+use std::collections::BTreeMap;
+
+use crate::{
+    hbase_controller::build_recommended_labels, zookeeper::ZookeeperConnectionInformation,
+};
+use stackable_hbase_crd::{HbaseCluster, HbaseRole, HBASE_SITE_XML};
 use stackable_operator::{
     builder::{ConfigMapBuilder, ObjectMetaBuilder},
     commons::product_image_selection::ResolvedProductImage,
     error::OperatorResult,
     k8s_openapi::api::core::v1::ConfigMap,
 };
-use std::collections::HashMap;
 
 /// Creates a discovery config map containing the `hbase-site.xml` for clients.
 pub fn build_discovery_configmap(
     hbase: &HbaseCluster,
-    zookeeper_connect_string: &str,
+    zookeeper_connection_information: &ZookeeperConnectionInformation,
     resolved_product_image: &ResolvedProductImage,
 ) -> OperatorResult<ConfigMap> {
-    let hbase_site_data: HashMap<String, Option<String>> = [(
-        HBASE_ZOOKEEPER_QUORUM.to_string(),
-        Some(zookeeper_connect_string.to_string()),
-    )]
-    .into();
+    let hbase_site = zookeeper_connection_information.as_hbase_settings();
 
     ConfigMapBuilder::new()
         .metadata(
@@ -35,7 +34,13 @@ pub fn build_discovery_configmap(
         )
         .add_data(
             HBASE_SITE_XML,
-            stackable_operator::product_config::writer::to_hadoop_xml(hbase_site_data.iter()),
+            stackable_operator::product_config::writer::to_hadoop_xml(
+                hbase_site
+                    .into_iter()
+                    .map(|(k, v)| (k, Some(v)))
+                    .collect::<BTreeMap<_, _>>()
+                    .iter(),
+            ),
         )
         .build()
 }