+++ rust/operator-binary/src/main.rs
@@ -1,6 +1,7 @@
 mod discovery;
 mod hbase_controller;
 mod product_logging;
+mod zookeeper;
 
 use crate::hbase_controller::HBASE_CONTROLLER_NAME;